+++ README.md
@@ -1,3 +1,3 @@
-# go-gin-template
+# dena-autumn-backend
 
 ハッカソンとかでサーバサイドを爆速開発するためのテンプレート
