+++ server/go.mod
@@ -1,8 +1,10 @@
-module github.com/shinbunbun/go-gin-template
+module github.com/shinbunbun/dena-autumn-backend/server
 
 go 1.20
 
 require github.com/gin-contrib/cors v1.4.0
+
+require github.com/jinzhu/inflection v1.0.0 // indirect
 
 require (
 	github.com/gin-contrib/sse v0.1.0 // indirect
@@ -10,7 +12,9 @@ require (
 	github.com/go-playground/locales v0.14.0 // indirect
 	github.com/go-playground/universal-translator v0.18.0 // indirect
 	github.com/go-playground/validator/v10 v10.10.0 // indirect
+	github.com/go-sql-driver/mysql v1.7.1
 	github.com/goccy/go-json v0.9.7 // indirect
+	github.com/jinzhu/gorm v1.9.16
 	github.com/json-iterator/go v1.1.12 // indirect
 	github.com/leodido/go-urn v1.2.1 // indirect
 	github.com/mattn/go-isatty v0.0.14 // indirect
