+++ server/model/ping.go
